delete courses

+++ backend/controllers/course.controller.js
@@ -5,6 +5,7 @@ import cloudinary from "../utils/cloudinary.js";
 
 export const createCourse = async (req, res) => {
   const { title, description, videoUrl, thumbnail } = req.body;
+  const instructorId = req.user.id;
   const file = req.files.thumbnail;
 
   try {
@@ -34,6 +35,7 @@ export const createCourse = async (req, res) => {
         public_id: result.public_id,
         url: result.secure_url,
       },
+      instructorId,
     });
 
     return res.status(201).json({
@@ -90,7 +92,26 @@ export const featuredCourses = async (req, res) => {
 export const deleteCourse = async (req, res) => {
   try {
     const courseId = req.params.id;
+    const instructorId = req.user.id;
+
+    const course = await Course.findById(courseId);
+
+    if (!course) {
+      return res.status(404).json({ message: "Course not found" });
+    }
+
+    // Check if the instructor is the owner of the course
+    if (course.instructorId.toString() !== instructorId) {
+      return res
+        .status(403)
+        .json({ message: "You do not have permission to delete this course" });
+    }
+
     await Course.findByIdAndDelete(courseId);
+
+    console.log(courseId);
+    console.log(instructorId);
+    console.log(course);
 
     res.status(200).json({ message: "Course deleted successfully" });
   } catch (error) {
